+++ src/tada-commands.adb
@@ -130,6 +130,20 @@ package body Tada.Commands is
       end if;
    end Parse_Profile;
 
+   function Parse_Force (Arguments : CL_Arguments.Argument_List.Vector)
+     return Boolean
+   is
+      Arguments_Count : constant Natural := Natural (Arguments.Length);
+   begin
+      if Arguments_Count < 2 then
+         return False;
+      elsif Arguments (2) = "--force" then
+         return True;
+      else
+         raise Parse_Error with "invalid option '" & Arguments (2) & "'";
+      end if;
+   end Parse_Force;
+
    function Parse (Arguments : CL_Arguments.Argument_List.Vector) return Command is
       Arguments_Count : constant Natural := Natural (Arguments.Length);
    begin
@@ -151,6 +165,9 @@ package body Tada.Commands is
             when Build =>
                return (Kind => Build,
                        Build_Profile => Parse_Profile (Arguments));
+            when Cache =>
+               return (Kind => Cache,
+                       Force => Parse_Force (Arguments));
             when Clean =>
                return (Kind => Clean);
             when Help =>
@@ -393,12 +410,28 @@ package body Tada.Commands is
       end if;
    end Execute_Clean;
 
+   procedure Execute_Cache (Cmd : Command) is
+      Tada_Manifest : constant Config.Manifest := Config.Read (Packages.Manifest_Name);
+      Tada_Package : constant Packages.Package_Info := Packages.Create
+        (Tada_Manifest.Sections ("package") ("name"),
+         Tada_Manifest.Sections ("package") ("version"));
+   begin
+      if Cmd.Force and then
+         Package_Cache.Is_Cached (Tada_Package)
+      then
+         Directories.Delete_Tree (Package_Cache.Package_Path (Tada_Package));
+      end if;
+
+      Package_Cache.Cache_Package (Directories.Current_Directory);
+   end Execute_Cache;
+
    procedure Execute_Help is
    begin
       Text_IO.Put_Line ("Usage: tada [command] [options]");
       Text_IO.New_Line;
       Text_IO.Put_Line ("Commands:");
       Text_IO.Put_Line ("    build [--profile <p>]               Compile the package");
+      Text_IO.Put_Line ("    cache                               Install package to the local cache");
       Text_IO.Put_Line ("    clean                               Remove build artifacts");
       Text_IO.Put_Line ("    help                                Show this message");
       Text_IO.Put_Line ("    init <name> [--exe|--lib]           Create a new package");
@@ -540,14 +573,14 @@ package body Tada.Commands is
       case Cmd.Kind is
          when Help | Init | Version =>
             null;
-         when Build | Clean | Install | Run | Test =>
+         when Build | Cache | Clean | Install | Run | Test =>
             if not In_Package_Root then
                raise Execute_Error with "could not find '" & Packages.Manifest_Name & "' in current directory";
             end if;
       end case;
 
       case Cmd.Kind is
-         when Clean | Help | Init | Version =>
+         when Cache | Clean | Help | Init | Version =>
             null;
          when Install =>
             if not Exec_On_Path ("curl") then
@@ -565,6 +598,7 @@ package body Tada.Commands is
 
       case Cmd.Kind is
          when Build => Execute_Build (Cmd);
+         when Cache => Execute_Cache (Cmd);
          when Clean => Execute_Clean;
          when Help => Execute_Help;
          when Init => Execute_Init (Cmd);