+++ tests/tada_commands_tests.ads
@@ -46,4 +46,10 @@ package Tada_Commands_Tests is
    procedure Test_Parse_Init_Name_Package (T : in out Test_Context);
 
    procedure Test_Parse_Install (T : in out Test_Context);
+
+   procedure Test_Parse_Cache (T : in out Test_Context);
+
+   procedure Test_Parse_Cache_Force (T : in out Test_Context);
+
+   procedure Test_Parse_Cache_Foo (T : in out Test_Context);
 end Tada_Commands_Tests;